Add files via upload
This is version 2 of our WAD Project, at which we made a couple of changes to the CSS code, making it look more attractive. We also added constraints to make sure that a user does not enter details multiple times. And, there is also a validation system for when students add their details.